backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
aiosqlite==0.19.0
httpx==0.25.0
playwright==1.40.0
asyncpg==0.29.0
